requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
pydantic==2.5.0
httpx==0.28.1
aiohttp>=3.9.0
aiofiles==23.2.1
supabase>=2.11.0
python-jose[cryptography]==3.3.0
google-generativeai>=0.3.0
sqlite-vec>=0.1.0
dodopayments>=0.22.0
